pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "oasees_sdk"
version = "0.5.4"
description = "Oasees SDK"
authors = [
    "Achilleas Economopoulos",
    "Andreas Oikonomakis",
]
readme = "README.md"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]


[tool.poetry.dependencies]
Click="*"
web3="6.19.0"
ipfshttpclient="0.8.0a2"
requests="*"
python-dotenv="*"
kubernetes="*"
requests_toolbelt="*"
soorgeon="*"
ploomber="*"
pipreqs="*"
jupytext="*"
ipylab="*"


[tool.poetry.plugins."console_scripts"]
oasees-sdk = "oasees_sdk.cli.cli:cli"
